Refactor to use a single role & for consistency
This refactors things so that site.pp looks up a single role for node
from hiera. It also makes the way a user's ssh key(s) get onto the
system consistent and more maintainable.

For nodes that had classes being pulled from hiera before, those classes
have been moved to a role and that role is now listed in the node's data
file.