Update downloads to use bind mount
This was previously doing cp -r --link, which creates a hard link for each individual file at the time the command is run, and the creates param meant it never got updated. This instead uses a bind mount to ensure downloads always points to the downloads directory inside artifacts, which should work for both nginx and rsync.